comments, and handling allocation for 0 bytes

+++ mymalloc.h
@@ -5,7 +5,7 @@
 #include <stddef.h>
 
 
-#define malloc(x) mymalloc(x,__FILE__,__LINE__)
+#define malloc(x) mymalloc((size_t)x,__FILE__,__LINE__)
 #define free(x) myfree(x,__FILE__,__LINE__)
 #define SizeOfEntry sizeof(entry)
 #define sizeOfArray 4096
